Split fetching and processing
